9- remove badly implemented functions zetakinit / zetak
the interface did not make sense (it is impossible to initialize for
Dedekind zeta without specifying a domain where the function is to be
evaluated).

The following replacement will be commited soon:

Closest equivalent to zetakinit:
  L = lfuninit(x^2+1, [c, w, h]);
to compute zeta_Q(i)(s) for |Re(s - c)| < w, |Im(s)| < h. Then
  lfun(L, s)
as an analog to zetak(). Or directly lfun(x^2+1, s) if a single value
is needed. [#368, #1647]