13- polinterpolate could create illegal objects [#1837] (test-case by Peter Bruin)
37- [libpari] varnmin, varnmax
There is an ambiguity with the notion of "evaluating a polynomial at t" when
the polynomial is constant. It could be
1) 'polcoeff(P,0)';
2) 'polcoeff(P,0)*t^0', where , t^0 is a shortcut for "1 in the ring t
belongs to", and this makes a difference when t = 'z or Mod(1,3) for
instance.

For the time being, the behaviour is inconsistent: mostly 1) [historical
behaviour], but sometimes 2). Should be fixed centrally in gsubst /
poleval.

The 'interpol' test now contains results that are now valid objects
(at least!), but inconsistent due to problems outside of polint scope.
Those are marked with a comment, as in

test([1], [x], z)      \\ expect [x, z], currently returns [x, x]

Further commits should fix gsubst / poleval and update the test results.