New libpari function mt_nbthreads.
Return the actual number of threads that would be launched if
mt_queue_start was called.

This is equal to mt_nbthreads, except if mt_is_parallel() is true, in
which case it is 1.

When evaluating the expression

? A;parfor(i=1,n,B,v,C);D

mt_is_parallel() is false at A and D but true at B and C.
mt_is_thread() si true only at B