92- addprimes() creates issues with nfdisc [wrong results]
because it breaks the assumption that a potential composite in a
"bounded factorization" has to come last. This is a proof of concept
through a correct implementation of absZ_factor_limit_strict, which
fixes the poldiscfactors issue.

Other functions may still be affected. Must now look for all occurrences of
Z_factor_limit and absZ_factor_limit and check.